requirements.txt
fastapi
uvicorn[standard]
openai>=1.44.0
python-dotenv
chromadb